ENGR00258733 WEIM-NOR: mtd: cfi_cmdset_0002: Do not release the mutex lock
The NOR(M29W256GL7AN6E) may suffers a write-buffer timeout during
the bonnie++/ubifs stress test. Micron has confirmed that it is
a silicon bug in the Nor.

This patch is just a workaround to fix this issue for this NOR.

With this patch, the read/write/erase will do in the synchronous way.

Signed-off-by: Huang Shijie <b32955@freescale.com>