mxc: gpu_viv: Don't do an identity check for every commit
Everytime the GPU is woken up to process something an Identity
check is done to verify the GPU is present and active.  This is
really heavy handed and definitely causes excessive latency in
the kernel driver that is not needed.

FIXME: Find a better way to verify the GPU is ready to process
commands.