dts: CBi/HB: fix spdif device naming
With the merge of all the patching the device name for spdif
got changed to a more human readable name.  This differs from
both the 3.10 kernel as well as the upstream and there are
already userspace scripts that rely on the different name.

This reverts back to the old name until we can coordinate a
name change.